az_automate_tidbcloud/requirements.txt
azure-functions>=1.15.0
requests
